Create basic scene
There's a player with basic controls that still needs fixing.
The first person control scripts are just copies of the third
person control scripts. They'll be fixed later.

There are basic cauldrons just to place some objects in the scene.

There are Fire shaders and materials but displaying
them requires the Shader Graph package. This is for
yet another later commit.