Fix ForestPack trees
The prefabs in ForestPack/Trees now work with
the Terrain tree painting tool. The fix was to
rotate these models at the mesh level to the Y-up
orientation.

It's fine for regular game objects to be a Z-up
FBX import because Unity recognizes that and
corrects the rotation automatically. But the
Terrain tool in particular just doesn't work with
that.