Fix quest null exception bug
Cause:
Some listeners were pointing to the uninstantiated
quest manager prefab, which caused some null
references to be accessed.

Fix:
Point to the quest manager in the scene instead of
the prefab from the assets folder.

It was tricky one to catch because in the Unity Editor,
the listener response target just says the same thing,
"questOneManager (UI One)", regardless of whether it was
from the scene or from the assets folder.