Create cooking animation and soup shader
The Cauldron prefab under Assets/Models/Cauldron
can now jump around with fruits to show that it's
cooking.

It also has a soup surface that ripples. The
material on this surface has lots of parameters
that are tweakable via the Inspector.

The dancing cauldron is in Assets/Scenes/AnimationPlayground.unity